requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
colorama==0.4.4
greenlet==1.1.2
iniconfig==1.1.1
numpy==1.21.2
packaging==21.0
pandas==1.3.3
pluggy==1.0.0
psycopg2==2.9.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.25
toml==0.10.2
